CogVM sources as per VMMaker.oscog-eem.269.
Fix bug in primitiveClone/cloneContext: that causes the copy to be a word short.
Use isPointerNonInt: and isContextNonInt: in a few places.

Implement unknownBytecode.  Send unknownBytecode to the activeContext on
unknown bytecode if the selector is in the specialObjectsArray.

Fix page size bug in platforms/Cross/vm/sqHeapMap.c.

Add a trap command to scripts/versionInfoPlist and the mac mvm's to allow ^C.

More description for unix builds on 64-bit linux.

Have the upload script mark the upload directory and contents as read-only.


git-svn-id: http://squeakvm.org/svn/squeak/branches/Cog@2696 fa1542d4-bde8-0310-ad64-8ed1123d492a